+++ api/organisation/index.ts
@@ -6,13 +6,9 @@ import { http } from '../lib/http'
 
 async function handleOrganisationsRoute(req: express.Request, res: express.Response, next: express.NextFunction) {
     try {
-        console.log('handleAddressRoute')
-        //const jwt = await generateToken()
-        //const TOKEN = 'Bearer ' + jwt
-        //console.log('jwt is ' + TOKEN)
         const response = await http.get('http://localhost:8090/v1/organisations')
         console.log(response.data)
-        res.send(response.data)
+        res.send(response.data.organisations)
     } catch (error) {
         console.error(error)
         const errReport = { apiError: error.data.message, apiStatusCode: error.status, message: 'Organsiationsroute error' }
